@@ -100,7 +100,9 @@ class StreamChatState extends State<StreamChat> {
             return Theme(
               data: materialTheme.copyWith(
                 primaryIconTheme: streamTheme.primaryIconTheme,
-                accentColor: streamTheme.colorTheme.accentPrimary,
+                colorScheme: materialTheme.colorScheme.copyWith(
+                  secondary: streamTheme.colorTheme.accentPrimary,
+                ),
               ),
               child: StreamChatCore(
                 client: client,
