@@ -6,8 +6,8 @@ import 'package:rxdart/rxdart.dart';
 import 'package:stream_chat/src/client/channel.dart';
 import 'package:stream_chat/src/client/retry_policy.dart';
 import 'package:stream_chat/src/core/error/error.dart';
-import 'package:stream_chat/src/event_type.dart';
 import 'package:stream_chat/src/core/models/message.dart';
+import 'package:stream_chat/src/event_type.dart';
 import 'package:stream_chat/stream_chat.dart';
 
 /// The retry queue associated to a channel
