@@ -106,9 +106,7 @@ class ChannelListPage extends StatelessWidget {
               [StreamChat.of(context).currentUser!.id],
             ),
             sort: const [SortOption('last_message_at')],
-            pagination: const PaginationParams(
-              limit: 20,
-            ),
+            limit: 20,
           ),
         ),
       );
