@@ -75,9 +75,7 @@ class ChannelListPage extends StatelessWidget {
             [StreamChat.of(context).currentUser!.id],
           ),
           sort: const [SortOption('last_message_at')],
-          pagination: const PaginationParams(
-            limit: 20,
-          ),
+          limit: 20,
           channelWidget: const ChannelPage(),
         ),
       ),
