@@ -57,7 +57,7 @@ import 'package:stream_chat_flutter_core/stream_chat_flutter_core.dart';
 /// [errorBuilder] must all be supplied and not null.
 class UserListCore extends StatefulWidget {
   /// Instantiate a new [UserListCore]
-  UserListCore({
+  const UserListCore({
     required this.errorBuilder,
     required this.emptyBuilder,
     required this.loadingBuilder,
@@ -66,16 +66,10 @@ class UserListCore extends StatefulWidget {
     this.filter = const Filter.empty(),
     this.sort,
     this.presence,
-    @Deprecated(
-      "'pagination' is deprecated and shouldn't be used. "
-      "This property is no longer used, Please use 'limit' instead",
-    )
-        this.pagination,
     this.groupAlphabetically = false,
     this.userListController,
-    int? limit,
-  })  : limit = limit ?? pagination?.limit ?? 30,
-        super(key: key);
+    this.limit = 30,
+  }) : super(key: key);
 
   /// A [UserListController] allows reloading and pagination.
   /// Use [UserListController.loadData] and [UserListController.paginateData]
@@ -97,8 +91,7 @@ class UserListCore extends StatefulWidget {
   /// The query filters to use.
   /// You can query on any of the custom fields you've defined on the [Channel].
   /// You can also filter other built-in channel fields.
-  // TODO: Make it non-nullable in a future breaking release
-  final Filter? filter;
+  final Filter filter;
 
   /// The sorting used for the channels matching the filters.
   /// Sorting is based on field and direction, multiple sorting options can be
@@ -108,16 +101,6 @@ class UserListCore extends StatefulWidget {
 
   /// If true you’ll receive user presence updates via the websocket events
   final bool? presence;
-
-  /// Pagination parameters
-  /// limit: the number of users to return (max is 30)
-  /// offset: the offset (max is 1000)
-  /// message_limit: how many messages should be included to each channel
-  @Deprecated(
-    "'pagination' is deprecated and shouldn't be used. "
-    "This property is no longer used, Please use 'limit' instead",
-  )
-  final PaginationParams? pagination;
 
   /// The amount of users requested per API call.
   final int limit;
