@@ -2,9 +2,9 @@
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 import 'package:stream_chat_localizations/src/stream_chat_localizations.dart';
-import 'package:flutter_test/flutter_test.dart';
 
 class FooStreamChatLocalizations extends StreamChatLocalizationsEn {
   FooStreamChatLocalizations(
