@@ -1,8 +1,7 @@
 // coverage:ignore-file
 import 'package:moor/moor_web.dart';
-import 'package:stream_chat_persistence/src/stream_chat_persistence_client.dart';
-
 import 'package:stream_chat_persistence/src/db/moor_chat_database.dart';
+import 'package:stream_chat_persistence/src/stream_chat_persistence_client.dart';
 
 /// A Helper class to construct new instances of [MoorChatDatabase] specifically
 /// for Web applications
