@@ -1,9 +1,9 @@
 import 'dart:math' as math;
 
+import 'package:stream_chat/stream_chat.dart';
 import 'package:stream_chat_persistence/src/dao/dao.dart';
 import 'package:stream_chat_persistence/src/db/moor_chat_database.dart';
 import 'package:test/test.dart';
-import 'package:stream_chat/stream_chat.dart';
 
 import '../../stream_chat_persistence_client_test.dart';
 
